+++ senscritique_export/row_utils/comics_utils.py
@@ -4,6 +4,8 @@ from typing import Dict, List
 
 from bs4 import element
 
+import senscritique_export.constants as cst
+
 from . import row_utils
 
 logger = logging.getLogger(__name__)
@@ -23,17 +25,17 @@ def get_comics_info_from_row(row: element.Tag) -> Dict:
         comics row info
     """
     return {
-        "Rank": row_utils.get_rank(row),
-        "Title": row_utils.get_title(row),
-        "URL": row_utils.get_url(row),
-        "Original Title": row_utils.get_original_title(row),
-        "Year": row_utils.get_year(row),
-        "Release Date": row_utils.get_baseline_0(row),
-        "Picture URL": row_utils.get_picture_url(row),
-        "Author": row_utils.get_producer(row),
-        "Description": row_utils.get_description(row),
-        "Average Rating": row_utils.get_average_rating(row),
-        "Number of Ratings": row_utils.get_number_of_ratings(row),
+        cst.RANK: row_utils.get_rank(row),
+        cst.TITLE: row_utils.get_title(row),
+        cst.URL: row_utils.get_url(row),
+        cst.ORIGINAL_TITLE: row_utils.get_original_title(row),
+        cst.YEAR: row_utils.get_year(row),
+        cst.RELEASE_DATE: row_utils.get_baseline_0(row),
+        cst.PICTURE_URL: row_utils.get_picture_url(row),
+        cst.AUTHOR: row_utils.get_producer(row),
+        cst.DESCRIPTION: row_utils.get_description(row),
+        cst.AVERAGE_RATING: row_utils.get_average_rating(row),
+        cst.NB_RATINGS: row_utils.get_number_of_ratings(row),
     }
 
 
@@ -46,15 +48,15 @@ def get_order_comics_columns() -> List:
        comic row columns order
     """
     return [
-        "Rank",
-        "Title",
-        "Author",
-        "Average Rating",
-        "Number of Ratings",
-        "URL",
-        "Original Title",
-        "Year",
-        "Release Date",
-        "Picture URL",
-        "Description",
+        cst.RANK,
+        cst.TITLE,
+        cst.AUTHOR,
+        cst.AVERAGE_RATING,
+        cst.NB_RATINGS,
+        cst.URL,
+        cst.ORIGINAL_TITLE,
+        cst.YEAR,
+        cst.RELEASE_DATE,
+        cst.PICTURE_URL,
+        cst.DESCRIPTION,
     ]