+++ senscritique_export/row_utils/movies_utils.py
@@ -4,6 +4,8 @@ from typing import Dict, List
 
 from bs4 import element
 
+import senscritique_export.constants as cst
+
 from . import row_utils
 
 logger = logging.getLogger(__name__)
@@ -23,19 +25,19 @@ def get_movies_info_from_row(row: element.Tag) -> Dict:
         movie row info
     """
     return {
-        "Rank": row_utils.get_rank(row),
-        "Title": row_utils.get_title(row),
-        "URL": row_utils.get_url(row),
-        "Original Title": row_utils.get_original_title(row),
-        "Year": row_utils.get_year(row),
-        "Release Date": row_utils.get_baseline_1(row),
-        "Length": row_utils.get_baseline_0(row),
-        "Picture URL": row_utils.get_picture_url(row),
-        "Genre": row_utils.get_genre(row),
-        "Producer": row_utils.get_producer(row),
-        "Description": row_utils.get_description(row),
-        "Average Rating": row_utils.get_average_rating(row),
-        "Number of Ratings": row_utils.get_number_of_ratings(row),
+        cst.RANK: row_utils.get_rank(row),
+        cst.TITLE: row_utils.get_title(row),
+        cst.URL: row_utils.get_url(row),
+        cst.ORIGINAL_TITLE: row_utils.get_original_title(row),
+        cst.YEAR: row_utils.get_year(row),
+        cst.RELEASE_DATE: row_utils.get_baseline_1(row),
+        cst.LENGTH: row_utils.get_baseline_0(row),
+        cst.PICTURE_URL: row_utils.get_picture_url(row),
+        cst.GENRE: row_utils.get_genre(row),
+        cst.PRODUCER: row_utils.get_producer(row),
+        cst.DESCRIPTION: row_utils.get_description(row),
+        cst.AVERAGE_RATING: row_utils.get_average_rating(row),
+        cst.NB_RATINGS: row_utils.get_number_of_ratings(row),
     }
 
 
@@ -48,17 +50,17 @@ def get_order_movies_columns() -> List:
         order of columns for movies rows
     """
     return [
-        "Rank",
-        "Title",
-        "Producer",
-        "Average Rating",
-        "Number of Ratings",
-        "URL",
-        "Original Title",
-        "Year",
-        "Release Date",
-        "Length",
-        "Picture URL",
-        "Genre",
-        "Description",
+        cst.RANK,
+        cst.TITLE,
+        cst.PRODUCER,
+        cst.AVERAGE_RATING,
+        cst.NB_RATINGS,
+        cst.URL,
+        cst.ORIGINAL_TITLE,
+        cst.YEAR,
+        cst.RELEASE_DATE,
+        cst.LENGTH,
+        cst.PICTURE_URL,
+        cst.GENRE,
+        cst.DESCRIPTION,
     ]