+++ senscritique_export/row_utils/music_utils.py
@@ -4,6 +4,8 @@ from typing import Dict, List
 
 from bs4 import element
 
+import senscritique_export.constants as cst
+
 from . import row_utils
 
 logger = logging.getLogger(__name__)
@@ -23,17 +25,17 @@ def get_music_info_from_row(row: element.Tag) -> Dict:
         music row info
     """
     return {
-        "Rank": row_utils.get_rank(row),
-        "Title": row_utils.get_title(row),
-        "URL": row_utils.get_url(row),
-        "Year": row_utils.get_year(row),
-        "Release Date": row_utils.get_baseline_0(row),
-        "Genre": row_utils.get_baseline_1(row),
-        "Number of Songs": row_utils.get_number_of_seasons(row),
-        "Picture URL": row_utils.get_picture_url(row),
-        "Artist": row_utils.get_producer(row),
-        "Average Rating": row_utils.get_average_rating(row),
-        "Number of Ratings": row_utils.get_number_of_ratings(row),
+        cst.RANK: row_utils.get_rank(row),
+        cst.TITLE: row_utils.get_title(row),
+        cst.URL: row_utils.get_url(row),
+        cst.YEAR: row_utils.get_year(row),
+        cst.RELEASE_DATE: row_utils.get_baseline_0(row),
+        cst.GENRE: row_utils.get_baseline_1(row),
+        cst.NB_SONGS: row_utils.get_number_of_seasons(row),
+        cst.PICTURE_URL: row_utils.get_picture_url(row),
+        cst.ARTIST: row_utils.get_producer(row),
+        cst.AVERAGE_RATING: row_utils.get_average_rating(row),
+        cst.NB_RATINGS: row_utils.get_number_of_ratings(row),
     }
 
 
@@ -46,15 +48,15 @@ def get_order_music_columns() -> List:
         order columns for music rows
     """
     return [
-        "Rank",
-        "Title",
-        "Average Rating",
-        "Number of Ratings",
-        "URL",
-        "Year",
-        "Release Date",
-        "Genre",
-        "Number of Songs",
-        "Picture URL",
-        "Artist",
+        cst.RANK,
+        cst.TITLE,
+        cst.AVERAGE_RATING,
+        cst.NB_RATINGS,
+        cst.URL,
+        cst.YEAR,
+        cst.RELEASE_DATE,
+        cst.GENRE,
+        cst.NB_SONGS,
+        cst.PICTURE_URL,
+        cst.ARTIST,
     ]