+++ tests/integration/test_senscritique_export.py
@@ -1,11 +1,12 @@
-"""Script used to get a collection from senscritique from a user name"""
 import pandas as pd
+import pytest
 
 from senscritique_export.scrapper import get_user_collection
 
-if __name__ == "__main__":
-    # Name of the user
-    user_name = "Jigot"
+
+@pytest.mark.parametrize("user_name", ["Jigot", "SubwaySam", "FS-10"])
+def test_get_user_colection(user_name):
 
     user_collection = get_user_collection(user=user_name)
     df_user_collection = pd.DataFrame(user_collection)
+    assert len(df_user_collection) > 1